+++ src/builtins/global_functions.rs
@@ -1,5 +1,6 @@
 use std::collections::HashMap;
 
+use chrono::prelude::*;
 use serde_json::value::{from_value, to_value, Value};
 
 use errors::Result;
@@ -13,7 +14,7 @@ pub fn make_range_fn() -> GlobalFn {
             Some(val) => match from_value::<usize>(val.clone()) {
                 Ok(v) => v,
                 Err(_) => bail!(
-                    "Global function `range` received start={} but `start` can only be a number"
+                    "Global function `range` received start={} but `start` can only be a number", val
                 ),
             },
             None => 0,
@@ -22,7 +23,7 @@ pub fn make_range_fn() -> GlobalFn {
             Some(val) => match from_value::<usize>(val.clone()) {
                 Ok(v) => v,
                 Err(_) => bail!(
-                    "Global function `range` received step_by={} but `step` can only be a number"
+                    "Global function `range` received step_by={} but `step` can only be a number", val
                 ),
             },
             None => 1,
@@ -31,7 +32,7 @@ pub fn make_range_fn() -> GlobalFn {
             Some(val) => match from_value::<usize>(val.clone()) {
                 Ok(v) => v,
                 Err(_) => {
-                    bail!("Global function `range` received end={} but `end` can only be a number")
+                    bail!("Global function `range` received end={} but `end` can only be a number", val)
                 }
             },
             None => bail!("Global function `range` was called without a `end` argument"),
@@ -51,16 +52,53 @@ pub fn make_range_fn() -> GlobalFn {
     })
 }
 
+pub fn make_now_fn() -> GlobalFn {
+    Box::new(move |args| -> Result<Value> {
+        let utc = match args.get("utc") {
+            Some(val) => match from_value::<bool>(val.clone()) {
+                Ok(v) => v,
+                Err(_) => bail!(
+                    "Global function `now` received utc={} but `utc` can only be a boolean", val
+                ),
+            },
+            None => false,
+        };
+        let timestamp = match args.get("timestamp") {
+            Some(val) => match from_value::<bool>(val.clone()) {
+                Ok(v) => v,
+                Err(_) => bail!(
+                    "Global function `now` received timestamp={} but `timestamp` can only be a boolean", val
+                ),
+            },
+            None => false,
+        };
+
+        if utc {
+            let datetime = Utc::now();
+            if timestamp {
+                return Ok(to_value(datetime.timestamp()).unwrap());
+            }
+            Ok(to_value(datetime.to_rfc3339()).unwrap())
+        } else {
+            let datetime = Local::now();
+            if timestamp {
+                return Ok(to_value(datetime.timestamp()).unwrap());
+            }
+            Ok(to_value(datetime.to_rfc3339()).unwrap())
+        }
+    })
+}
+
 #[cfg(test)]
 mod tests {
     use std::collections::HashMap;
 
     use serde_json::value::to_value;
 
-    use super::make_range_fn;
+    use super::*;
 
     #[test]
-    fn test_range_default() {
+    fn range_default() {
         let mut args = HashMap::new();
         args.insert("end".to_string(), to_value(5).unwrap());
 
@@ -69,7 +107,7 @@ mod tests {
     }
 
     #[test]
-    fn test_range_start() {
+    fn range_start() {
         let mut args = HashMap::new();
         args.insert("end".to_string(), to_value(5).unwrap());
         args.insert("start".to_string(), to_value(1).unwrap());
@@ -79,7 +117,7 @@ mod tests {
     }
 
     #[test]
-    fn test_range_start_greater_than_end() {
+    fn range_start_greater_than_end() {
         let mut args = HashMap::new();
         args.insert("end".to_string(), to_value(5).unwrap());
         args.insert("start".to_string(), to_value(6).unwrap());
@@ -88,12 +126,43 @@ mod tests {
     }
 
     #[test]
-    fn test_range_step_by() {
+    fn range_step_by() {
         let mut args = HashMap::new();
         args.insert("end".to_string(), to_value(10).unwrap());
         args.insert("step_by".to_string(), to_value(2).unwrap());
 
         let res = make_range_fn()(args).unwrap();
         assert_eq!(res, to_value(vec![0, 2, 4, 6, 8]).unwrap());
+    }
+
+    #[test]
+    fn now_default() {
+        let mut args = HashMap::new();
+
+        let res = make_now_fn()(args).unwrap();
+        assert!(res.is_string());
+        assert!(res.as_str().unwrap().contains("T"));
+    }
+
+    #[test]
+    fn now_datetime_utc() {
+        let mut args = HashMap::new();
+        args.insert("utc".to_string(), to_value(true).unwrap());
+
+        let res = make_now_fn()(args).unwrap();
+        assert!(res.is_string());
+        let val = res.as_str().unwrap();
+        println!("{}", val);
+        assert!(val.contains("T"));
+        assert!(val.contains("+00:00"));
+    }
+
+    #[test]
+    fn now_timestamp() {
+        let mut args = HashMap::new();
+        args.insert("timestamp".to_string(), to_value(true).unwrap());
+
+        let res = make_now_fn()(args).unwrap();
+        assert!(res.is_number());
     }
 }