+++ src/lib.rs
@@ -11,6 +11,7 @@
 extern crate glob;
 extern crate pest;
 extern crate serde;
+#[macro_use]
 extern crate serde_json;
 #[macro_use]
 extern crate pest_derive;