+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.11.8 (unreleased)
+
+- Add `True` and `False` as boolean values to match Python
+- Allow user to define their own escape function, if you want to generate JSON for example
+- Add `end` argument to the `truncate` filter to override the default ellipsis
+- Add a `group_by` filter
+- Add a `filter` filter
+- Add the `~` operator to concatenate strings
+- Add a `now` and `utc_now` global functions to get local and UTC datetimes
+- Add feature to enable the `preserve_order` feature of serde_json
+- Less confusing behaviour with math arithmetics
+
 ## 0.11.7 (2018-04-24)
 
 - Add array literal instantiation from inside Tera for set, set_global, kwargs